Rename utils.hpp to queue.hpp to reflect it's use.
The only thing inside utils.hpp is a thread safe queue, so name the file
accordingly.
