+++ tfg/audio_processor.hpp
@@ -10,7 +10,7 @@
 #include "../cpp/driver/everloop.h"
 #include "../cpp/driver/everloop_image.h"
 #include "mqtt/async_client.h"
-#include "utils.hpp"
+#include "queue.hpp"
 
 // Variables globales
 extern std::atomic<bool> running;
