+++ tfg/matrix_read.cpp
@@ -17,7 +17,7 @@
 #include "../cpp/driver/everloop_image.h"
 
 #include "audio_processor.hpp"
-#include "utils.hpp"
+#include "queue.hpp"
 
 // Configuración de MQTT
 const std::string SERVER_IP    = "tcp://localhost";
